Make use of LIBXS_UPDIV macro.

+++ src/libxs_malloc.c
@@ -530,7 +530,8 @@ LIBXS_API_INLINE size_t internal_get_scratch_size(const internal_malloc_pool_typ
 #if !defined(LIBXS_MALLOC_SCRATCH_MAX_NPOOLS) || (1 >= (LIBXS_MALLOC_SCRATCH_MAX_NPOOLS))
   LIBXS_UNUSED(exclude);
 #else
-  const internal_malloc_pool_type* pool = (const internal_malloc_pool_type*)LIBXS_UP2(internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
+  const internal_malloc_pool_type* pool = (const internal_malloc_pool_type*)LIBXS_UP2(
+    (uintptr_t)internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
 # if (1 < (LIBXS_MALLOC_SCRATCH_MAX_NPOOLS))
   const internal_malloc_pool_type *const end = pool + libxs_scratch_pools;
   LIBXS_ASSERT(libxs_scratch_pools <= LIBXS_MALLOC_SCRATCH_MAX_NPOOLS);
@@ -559,7 +560,8 @@ LIBXS_API_INLINE size_t internal_get_scratch_size(const internal_malloc_pool_typ
 LIBXS_API_INLINE internal_malloc_pool_type* internal_scratch_malloc_pool(const void* memory)
 {
   internal_malloc_pool_type* result = NULL;
-  internal_malloc_pool_type* pool = (internal_malloc_pool_type*)LIBXS_UP2(internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
+  internal_malloc_pool_type* pool = (internal_malloc_pool_type*)LIBXS_UP2(
+    (uintptr_t)internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
   const char* const buffer = (const char*)memory;
 #if defined(LIBXS_MALLOC_SCRATCH_MAX_NPOOLS) && (1 < (LIBXS_MALLOC_SCRATCH_MAX_NPOOLS))
   const unsigned int npools = libxs_scratch_pools;
@@ -657,7 +659,8 @@ LIBXS_API_INTERN void internal_scratch_malloc(void** memory, size_t size, size_t
     size_t local_size = 0;
 #if defined(LIBXS_MALLOC_SCRATCH_MAX_NPOOLS) && (0 < (LIBXS_MALLOC_SCRATCH_MAX_NPOOLS))
     if (0 < libxs_scratch_pools) {
-      internal_malloc_pool_type *const pools = (internal_malloc_pool_type*)LIBXS_UP2(internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
+      internal_malloc_pool_type *const pools = (internal_malloc_pool_type*)LIBXS_UP2(
+        (uintptr_t)internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
       internal_malloc_pool_type *const end = pools + libxs_scratch_pools, *pool = pools;
       const size_t align_size = libxs_alignment(size, alignment), alloc_size = size + align_size - 1;
 # if (0 != LIBXS_SYNC)
@@ -1886,8 +1889,10 @@ LIBXS_API_INTERN int libxs_xmalloc(void** memory, size_t size, size_t alignment,
 #endif
         /*fall-back*/NULL != buffer)
       {
-        char *const cbuffer = (char*)buffer, *const aligned = LIBXS_ALIGN(cbuffer + extra_size + sizeof(internal_malloc_info_type), alloc_alignment);
-        internal_malloc_info_type *const buffer_info = (internal_malloc_info_type*)(aligned - sizeof(internal_malloc_info_type));
+        char *const cbuffer = (char*)buffer, *const aligned = LIBXS_ALIGN(
+          cbuffer + extra_size + sizeof(internal_malloc_info_type), alloc_alignment);
+        internal_malloc_info_type *const buffer_info = (internal_malloc_info_type*)(
+          aligned - sizeof(internal_malloc_info_type));
         LIBXS_ASSERT((aligned + size) <= (cbuffer + alloc_size));
         LIBXS_ASSERT(0 < alloc_alignment);
         /* former content must be preserved prior to setup of buffer_info */
@@ -2337,7 +2342,8 @@ LIBXS_API_INTERN void libxs_xrelease_scratch(LIBXS_LOCK_TYPE(LIBXS_LOCK)* lock)
 # endif
   {
     unsigned int i;
-    pools = (internal_malloc_pool_type*)LIBXS_UP2(internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
+    pools = (internal_malloc_pool_type*)LIBXS_UP2(
+      (uintptr_t)internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
     for (i = 0; i < libxs_scratch_pools; ++i) {
       if (0 != pools[i].instance.minsize) {
         if (
@@ -2419,7 +2425,8 @@ LIBXS_API int libxs_get_scratch_info(libxs_scratch_info* info)
     info->internal = internal_malloc_private_max;
     info->local = internal_malloc_local_max;
     info->size = internal_malloc_public_max;
-    { const internal_malloc_pool_type* pool = (const internal_malloc_pool_type*)LIBXS_UP2(internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
+    { const internal_malloc_pool_type* pool = (const internal_malloc_pool_type*)LIBXS_UP2(
+        (uintptr_t)internal_malloc_pool_buffer, LIBXS_MALLOC_SCRATCH_PADDING);
 # if (1 < (LIBXS_MALLOC_SCRATCH_MAX_NPOOLS))
       const internal_malloc_pool_type *const end = pool + libxs_scratch_pools;
       LIBXS_ASSERT(libxs_scratch_pools <= LIBXS_MALLOC_SCRATCH_MAX_NPOOLS);
